Rename true-on-policy plugin package
