+++ megatron/core/extensions/sglang.py
@@ -1,6 +1,6 @@
 """Compatibility imports for the SGLang-compatible true-on-policy backend.
 
-New code should import from :mod:`miles_plugins.true_on_policy.sglang_backend`.
+New code should import from :mod:`miles_megatron_plugins.true_on_policy.sglang_backend`.
 """
 
-from miles_plugins.true_on_policy.sglang_backend import *  # noqa: F403
+from miles_megatron_plugins.true_on_policy.sglang_backend import *  # noqa: F403
