+++ megatron/core/tensor_parallel/matmul_tp_inv.py
@@ -1,6 +1,6 @@
 """Compatibility imports for true-on-policy TP-invariant matmul helpers.
 
-New code should import from :mod:`miles_plugins.true_on_policy.matmul`.
+New code should import from :mod:`miles_megatron_plugins.true_on_policy.matmul`.
 """
 
-from miles_plugins.true_on_policy.matmul import *  # noqa: F403
+from miles_megatron_plugins.true_on_policy.matmul import *  # noqa: F403
