+++ megatron/core/transformer/transformer_block.py
@@ -30,8 +30,8 @@ from megatron.core.transformer.transformer_layer import (
     get_transformer_layer_offset,
 )
 from megatron.core.transformer.utils import sharded_state_dict_default
-from miles_plugins.true_on_policy.contracts import resolve_true_on_policy_runtime_policy
-from miles_plugins.true_on_policy.sglang_backend import SGLangFinalRMSNorm, SGLangNorm
+from miles_megatron_plugins.true_on_policy.contracts import resolve_true_on_policy_runtime_policy
+from miles_megatron_plugins.true_on_policy.sglang_backend import SGLangFinalRMSNorm, SGLangNorm
 from megatron.core.utils import (
     WrappedTensor,
     deprecate_inference_params,
